+++ lib/project_types/theme/messages/messages.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 module Theme
   module Messages
     MESSAGES = {
@@ -18,8 +19,9 @@ module Theme
         ENSURE_USER
         stable_flag_suggestion: "If the current command isn't working as expected," \
           " we suggest re-running the command with the {{command: --stable}} flag",
-        confirm_current_directory: "It doesn’t seem like you’re running this command in a theme directory. " \
-          "Are you sure you want to proceed?",
+        current_directory_is_not_theme_directory: "It doesn’t seem like you’re running this command" \
+          " in a theme directory.",
+        confirm_current_directory: "Are you sure you want to proceed?",
         init: {
           help: <<~HELP,
             {{command:%s theme init}}: Clones a Git repository to use as a starting point for building a new theme.